Address deprecations from persistence
A backwards-compatibility layer has been added to persistence to help
consumers move to the new namespacing. It is based on class aliases,
which means the type declaration changes should not be a BC-break: types
are the same.
See doctrine/persistence#71

This means:
- using the new namespaces
- adding autoload calls for new types to types that may be extended and
use persistence types in type declarations of non-constructor methods,
so that signature compatibility is recognized by old versions of php.
Luckily, there are none of those in this bundle.

This explicitely requires the persistence lib, which was an
unworded dependency of this bundle. I think it was not added because it
the depency was not noticed, but there might be more implicit
dependencies (doctrine/common, doctrine/doctrine-bundle).

Addressing the deprecations fixes the unit tests, because one of them
was expecting the old namespace in an exception message.